+++ services/gateway/apis/http-version1-compat/methods/accountVotes.js
@@ -14,7 +14,7 @@
  *
  */
 const votesSource = require('../../../sources/version1/votes');
-const envelope = require('../../../sources/mappings/stdEnvelope');
+const envelope = require('../../../sources/version1/mappings/stdEnvelope');
 
 module.exports = {
 	version: '2.0',
