+++ services/gateway/apis/http-version1-compat/methods/accountsTop.js
@@ -14,7 +14,7 @@
  *
  */
 const accountsSource = require('../../../sources/version1/accounts');
-const envelope = require('../../../sources/mappings/stdEnvelope');
+const envelope = require('../../../sources/version1/mappings/stdEnvelope');
 
 module.exports = {
 	version: '2.0',
