+++ services/gateway/apis/http-version1-compat/methods/block.js
@@ -14,7 +14,7 @@
  *
  */
 const blocksSource = require('../../../sources/version1/blocks');
-const envelope = require('../../../sources/mappings/stdEnvelope');
+const envelope = require('../../../sources/version1/mappings/stdEnvelope');
 
 module.exports = {
 	version: '2.0',
