+++ services/gateway/apis/http-version1-compat/methods/voters.js
@@ -14,7 +14,7 @@
  *
  */
 const votersSource = require('../../../sources/version1/voters');
-const envelope = require('../../../sources/mappings/stdEnvelope');
+const envelope = require('../../../sources/version1/mappings/stdEnvelope');
 const { transformParams, response } = require('../../http-version1/swagger/utils');
 
 module.exports = {
