+++ services/gateway/apis/http-version1/methods/accounts.js
@@ -14,7 +14,7 @@
  *
  */
 const accountsSource = require('../../../sources/version1/accounts');
-const envelope = require('../../../sources/mappings/stdEnvelope');
+const envelope = require('../../../sources/version1/mappings/stdEnvelope');
 const { transformParams, response } = require('../swagger/utils');
 
 module.exports = {
