+++ services/gateway/apis/http-version1/methods/delegatesNextForgers.js
@@ -14,7 +14,7 @@
  *
  */
 const delegatesSource = require('../../../sources/version1/delegates');
-const envelope = require('../../../sources/mappings/stdEnvelope');
+const envelope = require('../../../sources/version1/mappings/stdEnvelope');
 const { transformParams, response } = require('../swagger/utils');
 
 module.exports = {
