+++ services/gateway/apis/http-version1/methods/networkStatistics.js
@@ -14,7 +14,7 @@
  *
  */
 const networkStatisticsSource = require('../../../sources/version1/networkStatistics');
-const envelope = require('../../../sources/mappings/stdEnvelope');
+const envelope = require('../../../sources/version1/mappings/stdEnvelope');
 
 module.exports = {
 	version: '2.0',
