+++ services/gateway/apis/http-version1/methods/peers.js
@@ -14,7 +14,7 @@
  *
  */
 const peersSource = require('../../../sources/version1/peers');
-const envelope = require('../../../sources/mappings/stdEnvelope');
+const envelope = require('../../../sources/version1/mappings/stdEnvelope');
 const { transformParams, response } = require('../swagger/utils');
 
 module.exports = {
