+++ services/gateway/apis/http-version1/methods/peersDisconnected.js
@@ -14,7 +14,7 @@
  *
  */
 const peersSource = require('../../../sources/version1/peers');
-const envelope = require('../../../sources/mappings/stdEnvelope');
+const envelope = require('../../../sources/version1/mappings/stdEnvelope');
 
 module.exports = {
 	version: '2.0',
