+++ services/gateway/apis/http-version1/methods/transactions.js
@@ -14,7 +14,7 @@
  *
  */
 const transactionsSource = require('../../../sources/version1/transactions');
-const envelope = require('../../../sources/mappings/stdEnvelope');
+const envelope = require('../../../sources/version1/mappings/stdEnvelope');
 const { transformParams, response } = require('../swagger/utils');
 
 module.exports = {
