+++ services/gateway/apis/http-version1/methods/transactionsStatisticsMonth.js
@@ -14,7 +14,7 @@
  *
  */
 const transactionsStatisticsMonthSource = require('../../../sources/version1/transactionsStatisticsMonth');
-const envelope = require('../../../sources/mappings/stdEnvelope');
+const envelope = require('../../../sources/version1/mappings/stdEnvelope');
 const { transformParams, response } = require('../swagger/utils');
 
 module.exports = {
