+++ services/gateway/apis/http-version2/methods/blocks.js
@@ -14,7 +14,7 @@
  *
  */
 const blocksSource = require('../../../sources/version2/blocks');
-const envelope = require('../../../sources/mappings/stdEnvelope');
+const envelope = require('../../../sources/version2/mappings/stdEnvelope');
 
 module.exports = {
 	version: '2.0',
