+++ services/gateway/sources/version1/votes.js
@@ -13,7 +13,7 @@
  * Removal or modification of this copyright notice is prohibited.
  *
  */
-const vote = require('../mappings/vote');
+const vote = require('./mappings/vote');
 
 module.exports = {
 	type: 'moleculer',
